move PyItem and PyList classes to the plugin
