+++ examples/todo-pylist.html
@@ -40,6 +40,7 @@
 
                 <py-script>
                     from js import document
+                    from datetime import datetime as dt
                     from pyodide.ffi.wrappers import add_event_listener
 
                     def add_task(*args, **kws):
